Refactor imports to use absolute paths for components and styles across coding test and community pages
