+++ frontend/src/app/coding-test/progress/page.tsx
@@ -1,12 +1,12 @@
 "use client";
 import React, { useState, useEffect, useRef } from "react";
 import Head from "next/head";
-import Header from "../../../components/header";
-import Footer from "../../../components/Footer";
-import CodeEditor from "../../../components/CodeEditor";
+import Header from "@/components/Header";
+import Footer from "@/components/Footer";
+import CodeEditor from "@/components/CodeEditor";
 import Link from "next/link";
 import { useRouter, useSearchParams } from "next/navigation";
-import styles from "../../../styles/coding-test.module.css";
+import styles from "@/styles/coding-test.module.css";
 
 const CodingTestProgressPage: React.FC = () => {
   const router = useRouter();
