+++ frontend/src/app/community/create/page.tsx
@@ -1,12 +1,12 @@
 "use client";
 import React, { useState, useEffect } from "react";
 import Head from "next/head";
-import Header from "../../../components/header";
-import Footer from "../../../components/Footer";
+import Header from "@/components/Header";
+import Footer from "@/components/Footer";
 import Link from "next/link";
 import { useRouter, useSearchParams } from "next/navigation";
-import CodeEditor from "../../../components/CodeEditor";
-import styles from "../../../styles/community.module.css";
+import CodeEditor from "@/components/CodeEditor";
+import styles from "@/styles/community.module.css";
 
 const CommunityCreatePage: React.FC = () => {
   const router = useRouter();
